requirements.txt
asgiref==3.5.1
asttokens==2.1.0
backcall==0.2.0
black==23.1.0
cfgv==3.3.1
click==8.1.3
decorator==5.1.1
distlib==0.3.6
Django==4.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
filelock==3.9.0
flake8==6.0.0
identify==2.5.18
ipython==8.6.0
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.22.3
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==3.0.0
pre-commit==3.1.0
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.13.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.1
tomli==2.0.1
traitlets==5.5.0
tzdata==2022.1
virtualenv==20.19.0
wcwidth==0.2.5
